Match GunshipA_Init(WORLDINFO_s*)
The function drives the shared mini-snowtrooper globals (trooper_boltid,
trooper_side, LevGizmo), not level-local statics, and its gizmo names are
the power_a/b pylons. Widening trooper_boltid/trooper_side to their real
sizes and correcting BoltType_FindIDByName's return width to i32 (the
original returns a full 32-bit index or -1) also drops the spurious cbw
from AnakinsFlightB_Init.

+++ src/gameapi/edtools/edstubs.cpp
@@ -72,7 +72,7 @@ extern "C" {
 
 } // extern "C"
 
-i8 BoltType_FindIDByName(char *name, WORLDINFO *world) {
+i32 BoltType_FindIDByName(char *name, WORLDINFO *world) {
     (void)name;
     (void)world;
     UNIMPLEMENTED();
